+++ src/main/kotlin/com/atlassian/performance/tools/jirasoftwareactions/api/actions/BrowseBoardsAction.kt
@@ -11,12 +11,34 @@ import net.jcip.annotations.NotThreadSafe
 class BrowseBoardsAction(
     private val jiraSoftware: WebJiraSoftware,
     private val meter: ActionMeter,
-    private val boardsMemory: AgileBoardIdMemory
+    private val boardsMemory: AgileBoardIdMemory,
+    private val scrumBoardsMemory: AgileBoardIdMemory
 ) : Action {
+
+    constructor(
+        jiraSoftware: WebJiraSoftware,
+        meter: ActionMeter,
+        boardsMemory: AgileBoardIdMemory
+    ) : this(
+        jiraSoftware = jiraSoftware,
+        meter = meter,
+        boardsMemory = boardsMemory,
+        scrumBoardsMemory = object : AgileBoardIdMemory {
+            override fun recall(): String? {
+                return null
+            }
+
+            override fun remember(memories: Collection<String>) {
+
+            }
+        }
+    )
+
     override fun run() {
         val browseBoardsPage =
             meter.measure(BROWSE_BOARDS) { jiraSoftware.goToBrowseBoards().waitForBoardsList() }
 
         boardsMemory.remember(browseBoardsPage.getBoardIds())
+        scrumBoardsMemory.remember(browseBoardsPage.getScrumBoardIds())
     }
 }